refactor: gracefully shutdown in case of manager stop

+++ tap-agent/src/main.rs
@@ -2,8 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use anyhow::Result;
+use ractor::ActorStatus;
 use tokio::signal::unix::{signal, SignalKind};
-use tracing::{debug, info};
+use tracing::{debug, error, info};
 
 use indexer_tap_agent::{agent, CONFIG};
 
@@ -13,13 +14,14 @@ async fn main() -> Result<()> {
     lazy_static::initialize(&CONFIG);
     debug!("Config: {:?}", *CONFIG);
 
-    let manager = agent::start_agent().await;
+    let (manager, handler) = agent::start_agent().await;
     info!("TAP Agent started.");
 
     // Have tokio wait for SIGTERM or SIGINT.
     let mut signal_sigint = signal(SignalKind::interrupt())?;
     let mut signal_sigterm = signal(SignalKind::terminate())?;
     tokio::select! {
+        _ = handler => error!("SenderAccountsManager stopped"),
         _ = signal_sigint.recv() => debug!("Received SIGINT."),
         _ = signal_sigterm.recv() => debug!("Received SIGTERM."),
     }
@@ -27,10 +29,12 @@ async fn main() -> Result<()> {
     info!("Shutting down...");
 
     // We don't want our actor to run any shutdown logic, so we kill it.
-    manager
-        .kill_and_wait(None)
-        .await
-        .expect("Failed to kill manager.");
+    if manager.get_status() == ActorStatus::Running {
+        manager
+            .kill_and_wait(None)
+            .await
+            .expect("Failed to kill manager.");
+    }
 
     // Stop the server and wait for it to finish gracefully.
     debug!("Goodbye!");
